Rename simply to store instead
